Creating a new dashboard component and moving upload form into it.
