+++ src/ui/app.jsx.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Module from '../assignments/project-1/index.jsx'
+import Gallery from '../assignments/project-1/index.jsx'
 
 const App = ({auth, ...props}) => {
 	switch (auth.status) {
 		case `init`: return <span>Authorizing</span>
 		case `failure`: return <span>{auth.message}</span>
 
-		default: return	<Module {...props} />
+		default: return	<Gallery {...props} />
 	}
 }
