+++ src/ui/components/uploader.jsx.js
@@ -17,7 +17,11 @@ class Uploader extends React.Component {
 
 	render() {
 		const {upload, ...inputProps} = this.props
-		return <input {...inputProps} multiple type="file" onChange={this.handleFiles} />
+		return <div>
+					<label tabindex="0" for="uploader" className="uploader">Upload Files
+						<input {...inputProps} className="uploader-input" id="uploader" multiple type="file" onChange={this.handleFiles} />
+					</label>
+				</div>
 	}
 
 }
